Lesson 11 - Access Control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,16 @@
     class PostsController extends Controller
     {
         /**
+         * Create a new controller instance.
+         *
+         * @return void
+         */
+        public function __construct()
+        {
+            $this->middleware('auth', ['except' => ['index', 'show']]);
+        }
+
+        /**
          * Display a listing of the resource.
          *
          * @return \Illuminate\Http\Response
@@ -75,6 +85,13 @@
         public function edit($id)
         {
             $post = Post::find($id);
+
+            // Check for correct user
+            if(auth()->user()->id != $post->user_id)
+            {
+                return redirect('/posts')->with('error', 'Unauthorised');
+            }
+
             return view('posts.edit')->with('post', $post);
         }
 
@@ -110,6 +127,13 @@
         public function destroy($id)
         {
             $post = Post::find($id);
+
+            // Check for correct user
+            if(auth()->user()->id != $post->user_id)
+            {
+                return redirect('/posts')->with('error', 'Unauthorised');
+            }
+
             $post->delete();
 
             return redirect('/posts')->with('success', 'Post Removed');
